fix(hermia-97t): judge_score NUMERIC not INTEGER per Gemini review
LLM-as-judge may produce float scores (averages, probability-based).
NUMERIC avoids truncation and matches consistency_pct pattern in schema.

+++ scripts/add_judge_columns.sql
@@ -4,5 +4,5 @@
 -- v0.3 LLM-as-judge work populates them via the --judge flag.
 
 ALTER TABLE hermia_results
-    ADD COLUMN IF NOT EXISTS judge_score     INTEGER,
+    ADD COLUMN IF NOT EXISTS judge_score     NUMERIC,
     ADD COLUMN IF NOT EXISTS judge_reasoning TEXT;
